Add support for building on MacOS X
This commit adds support for building on MacOS X. There are
two main changes needed for os x.

  * Use a .dylib extension for building, cleaning and installation
  * Use libtool to record dylib name with @rpath/ prefix

The dylib extension was a standard change for os x.

The libtool change is needed for the difference in dynamic
linking behavior on os x.  OS X differs from linux in that
it the linker embeds the full path to the shared library
when it records the dependency. Therefore to succesfully use
-rpath to find the library, the library must say that its
name should be found at @rpath/libintel_aes.dylib. The
@rpath will be expanded by dyld at runtime to the full path,
and an executable can link with a -rpath option to specify
the search path.